Move PPI to new layer file
