+++ scapy/layers/dot11.py
@@ -14,7 +14,6 @@
 # along with Scapy. If not, see <http://www.gnu.org/licenses/>.
 
 # Copyright (C) Philippe Biondi <phil@secdev.org>
-# Original PPI author: <jellch@harris.com>
 
 """
 Wireless LAN according to IEEE 802.11.
@@ -216,74 +215,6 @@ class RadioTap(Packet):
         if self.len is None:
             p = p[:2] + struct.pack("!H", len(p))[::-1] + p[3:]
         return p + pay
-
-
-# Dictionary to map the TLV type to the class name of a sub-packet
-_ppi_types = {}
-
-
-def addPPIType(id, value):
-    _ppi_types[id] = value
-
-
-def getPPIType(id, default="default"):
-    return _ppi_types.get(id, _ppi_types.get(default, None))
-
-
-# Default PPI Field Header
-class PPIGenericFldHdr(Packet):
-    name = "PPI Field Header"
-    fields_desc = [LEShortField('pfh_type', 0),
-                   FieldLenField('pfh_length', None, length_of="value", fmt='<H', adjust=lambda p, x:x + 4),
-                   StrLenField("value", "", length_from=lambda p:p.pfh_length)]
-
-    def extract_padding(self, p):
-        return b"", p
-
-
-def _PPIGuessPayloadClass(p, **kargs):
-    """ This function tells the PacketListField how it should extract the
-        TLVs from the payload.  We pass cls only the length string
-        pfh_len says it needs.  If a payload is returned, that means
-        part of the sting was unused.  This converts to a Raw layer, and
-        the remainder of p is added as Raw's payload.  If there is no
-        payload, the remainder of p is added as out's payload.
-    """
-    if len(p) >= 4:
-        t, pfh_len = struct.unpack("<HH", p[:4])
-        # Find out if the value t is in the dict _ppi_types.
-        # If not, return the default TLV class
-        cls = getPPIType(t, "default")
-        pfh_len += 4
-        out = cls(p[:pfh_len], **kargs)
-        if (out.payload):
-            out.payload = conf.raw_layer(out.payload.load)
-            out.payload.underlayer = out
-            if (len(p) > pfh_len):
-                out.payload.payload = conf.padding_layer(p[pfh_len:])
-                out.payload.payload.underlayer = out.payload
-        elif (len(p) > pfh_len):
-            out.payload = conf.padding_layer(p[pfh_len:])
-            out.payload.underlayer = out
-    else:
-        out = conf.raw_layer(p, **kargs)
-    return out
-
-
-class PPI(Packet):
-    name = "Per-Packet Information header (PPI)"
-    fields_desc = [ByteField('version', 0),
-                   ByteField('flags', 0),
-                   FieldLenField('len', None, length_of="PPIFieldHeaders", fmt="<H", adjust=lambda p, x: x + 8),
-                   LEIntField('dlt', 1),
-                   PacketListField("PPIFieldHeaders", [], _PPIGuessPayloadClass, length_from=lambda p: p.len - 8,)]
-
-    def guess_payload_class(self, payload):
-        return conf.l2types.get(self.dlt, Packet.guess_payload_class(self, payload))
-
-
-# Register PPI
-addPPIType("default", PPIGenericFldHdr)
 
 
 class Dot11(Packet):
@@ -764,8 +695,6 @@ class Dot11Ack(Packet):
 
 
 bind_layers(PrismHeader, Dot11,)
-bind_layers(PPI, Dot11, dlt=DLT_IEEE802_11)
-bind_layers(PPI, Ether, dlt=DLT_EN10MB)
 bind_layers(Dot11, LLC, type=2)
 bind_layers(Dot11QoS, LLC,)
 bind_layers(Dot11, Dot11AssoReq, subtype=0, type=0)
@@ -797,7 +726,6 @@ conf.l2types.register(DLT_PRISM_HEADER, PrismHeader)
 conf.l2types.register_num2layer(802, PrismHeader)
 conf.l2types.register(DLT_IEEE802_11_RADIO, RadioTap)
 conf.l2types.register_num2layer(803, RadioTap)
-conf.l2types.register(DLT_PPI, PPI)
 
 
 class WiFi_am(AnsweringMachine):
